refactor: refactor system tests don't require separate test build step standardize filenames to match unit tests add pytest configuration to pyproject.toml
