+++ pyproject.toml
@@ -64,7 +64,8 @@ dnnv_manage = "dnnv._manage.__main__:_main"
 
 [tool.coverage.run]
 branch = true
-command_line = "-m pytest tests/unit_tests tests/system_tests --ff --verbose --import-mode=importlib"
+# command_line = "-m pytest tests/unit_tests tests/system_tests --ff --verbose --import-mode=importlib"
+command_line = "-m pytest"
 parallel = true
 source = ["dnnv"]
 omit = [
@@ -73,3 +74,8 @@ omit = [
     "dnnv/logging_utils/*",
     "dnnv/_manage/*",
 ]
+
+[tool.pytest.ini_options]
+minversion = "6.0"
+addopts = "--ff --verbose --import-mode=importlib"
+testpaths = ["tests"]
