+++ tests/system_tests/conftest.py
@@ -1,11 +1,9 @@
-import pytest
 import sys
 
 from pathlib import Path
 
 sys.path.append(str(Path(__file__).parent.parent))
 
+from system_tests.artifacts.build_artifacts import build
 
-def pytest_collect_file(parent, path):
-    if path.basename.endswith("_tests.py"):
-        return pytest.Module.from_parent(parent, fspath=path)
+build()
