+++ tests/system_tests/test_verifiers/test_mnist.py
