+++ tests/system_tests/test_verifiers/test_neurify.py
@@ -2,7 +2,7 @@ import unittest
 
 from dnnv.verifiers.neurify import Neurify
 
-from system_tests.verifiers_tests.utils import VerifierTests
+from system_tests.test_verifiers.utils import VerifierTests
 
 
 @unittest.skipIf(not Neurify.is_installed(), "Neurify is not installed")
