+++ tests/system_tests/test_verifiers/test_random.py
@@ -36,7 +36,8 @@ VERIFIERS = {
 
 
 @unittest.skipIf(
-    sum(v.is_installed() for v in VERIFIERS.values()) < 2, "Not enough verifiers installed"
+    sum(v.is_installed() for v in VERIFIERS.values()) < 2,
+    "Not enough verifiers installed",
 )
 class RandomTests(unittest.TestCase):
     def setUp(self):
