+++ tests/system_tests/test_verifiers/test_verinet.py
@@ -2,7 +2,7 @@ import unittest
 
 from dnnv.verifiers.verinet import VeriNet
 
-from system_tests.verifiers_tests.utils import VerifierTests
+from system_tests.test_verifiers.utils import VerifierTests
 
 
 @unittest.skipIf(not VeriNet.is_installed(), "VeriNet is not installed")
